Implementacija/GameHub/requirements.txt
django
mysqlclient
channels[daphne]
channels_redis
pillow
selenium
